Added Record class, and have models returing an array of Records instead of a pure array.
Also overloaded the Record and Model classes.

	1. Calling a Record->member will first look for the same member in the model.  If it can't find it there it then;
	2. Looks for a function in the parent model wit the same name and calls that.  If it still can't find it there, it returns false.
	3. Calling a Record->function() will call the same function in the parent model (if it exists).

Thus, each record object has the same functionality as its parent Model.  And you'll conveniently have access to Model functions in the views.

And inside your models, you can use $this->field wherever you want.  $this->member will try to find the object var 1st, and if it's not found it will look inside that object for a function with the same name and call that one.